+++ antcrew/__init__.py
@@ -18,6 +18,7 @@ from antcrew.sandbox import DockerRunner, LocalRunner, SandboxRunner
 from antcrew.sandbox import RunResult as SandboxRunResult
 from antcrew.core.run_result import RunResult
 from antcrew.checkpointers import SqliteSaver
+from antcrew.core.exceptions import CostLimitExceeded
 from antcrew.flow import load_flow, validate_flow, format_flow
 from antcrew.project import Project
 from antcrew.config import load_context, TeamContext, build_llm, build_runner
@@ -73,6 +74,7 @@ __all__ = [
     "RunResult",
     "SandboxRunResult",
     "SqliteSaver",
+    "CostLimitExceeded",
     # Flow
     "load_flow",
     "validate_flow",